pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "transformers.*",
    "datasets.*",
    "numpy.*",
    "pandas.*",
    "matplotlib.*",
    "seaborn.*",
    "sklearn.*",
    "nltk.*",
    "wordcloud.*",
    "tqdm.*",
    "yaml.*",
    "dotenv.*",
    "wandb.*",
]
ignore_missing_imports = true

[project]
name = "imdb_bert_project"
version = "0.1.0"
description = "BERT-based sentiment analysis on IMDB Movie Review Dataset"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "numpy>=1.19.0",
    "pandas>=1.1.0",
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
    "scikit-learn>=0.24.0",
    "torch>=1.8.0",
    "transformers>=4.5.0",
    "datasets>=1.5.0",
    "nltk>=3.6.0",
    "wordcloud>=1.8.0",
    "tqdm>=4.50.0",
    "pyyaml>=5.4.0",
    "python-dotenv>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "black>=21.5b0",
    "isort>=5.8.0",
    "flake8>=3.9.0",
    "pytest>=6.2.0",
    "mypy>=0.812",
]
notebook = [
    "jupyter>=1.0.0",
    "notebook>=6.3.0",
]
tracking = [
    "wandb>=0.12.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/imdb_bert_project"
"Bug Tracker" = "https://github.com/yourusername/imdb_bert_project/issues"
